change data type of contribution count from int to long

+++ app-server/subprojects/bounded_context/challenge/application/src/main/kotlin/club/staircrusher/challenge/application/port/in/use_case/UpdateChallengeRanksUseCase.kt
@@ -45,14 +45,14 @@ class UpdateChallengeRanksUseCase(
                             id = EntityIdGenerator.generateRandom(),
                             challengeId = challenge.id,
                             userId = userId,
-                            contributionCount = contributions.size,
+                            contributionCount = contributions.size.toLong(),
                             rank = -1,
                             createdAt = now,
                             updatedAt = now,
                         )
 
                         challengeRank.apply {
-                            contributionCount = contributions.size
+                            contributionCount = contributions.size.toLong()
                             updatedAt = now
                         }
                     }
